% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingTemplates.R
\docType{package}
\name{testingTemplates}
\alias{testingTemplates}
\alias{testingTemplates-package}
\title{This Package Helps Illustrate Template Tricks}
\description{
Here we have collected a set of tricks or pieces
of code which we have found useful in formatting
output reports in cleaner, more standardized ways.
}
